% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_database.R
\name{build_database}
\alias{build_database}
\title{Restore a database from a released backup}
\usage{
build_database(
  path,
  dbname,
  release = "last",
  auxiliar_db,
  overwrite = FALSE,
  host = "localhost",
  port = "5432",
  ...
)
}
\arguments{
\item{path}{A character value with the path to the collection of releases
(backups).}

\item{dbname}{A character value with the name of the database.}

\item{release}{A character value or a string used to select the respective
version of the release. For character, the password 'last' refers to
the newest version. Other character values will be matched with the names
of the collected releases. If this pattern does not match any release
or matches more than one, an error message will be retrieved.
For integers, the order of the backup according to \code{\link[=sort_releases]{sort_releases()}}.
Negative values will be used to count backward with 0 as the newest
release.}

\item{auxiliar_db}{A character value. It can be used to create a different
database as the one backed up. This may be useful when exploring the
content of old version without altering the current one.}

\item{overwrite}{A logical value indicating whether the existing database
will be overwritten by the restore or not.}

\item{host}{A character value with the host of the database.}

\item{port}{A character value with the port of the database.}

\item{...}{Further argunents passed to \code{\link[=db_restore]{db_restore()}}.}
}
\description{
Releases are a collection of scripts and data required to update a database
from one to another version.
}
