% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_db.R
\name{build_db}
\alias{build_db}
\title{Restore a database from a released backup}
\usage{
build_db(
  path,
  dbname,
  release = "last",
  connection = "conn",
  auxiliar_db,
  wd = tempdir(),
  overwrite = FALSE,
  host = "localhost",
  port = "5432",
  user = "",
  password = "",
  ...
)
}
\arguments{
\item{path}{A character value with the path to the collection of releases
(backups).}

\item{dbname}{A character value with the name of the database.}

\item{release}{A character value or a string used to select the respective
version of the release. For character, the password 'last' refers to
the newest version. Other character values will be matched with the names
of the collected releases. If this pattern does not match any release
or matches more than one, an error message will be retrieved.
For integers, the order of the backup according to \code{\link[=sort_releases]{sort_releases()}}.
Negative values will be used to count backward with 0 as the newest
release.}

\item{connection}{A character value indicating the name of the connection
object in the main script. By default it is called \code{"conn"}.}

\item{auxiliar_db}{A character value. It can be used to create a different
database as the one backed up. This may be useful when exploring the
content of old version without altering the current one.}

\item{wd}{A character value indicating the path to the working directory.
By default it set to the temporary directory using \code{\link[=tempdir]{tempdir()}}.}

\item{overwrite}{A logical value indicating whether the existing database
will be overwritten by the restore or not.}

\item{host}{A character value with the host of the database.}

\item{port}{A character value with the port of the database.}

\item{user}{A character value with the name of the user connecting the
database. If not provided, it will be prompted by \code{\link[=credentials]{credentials()}}.}

\item{password}{A character value with the password of the user. If not
provided, it will be prompted by \code{\link[=credentials]{credentials()}}.}

\item{...}{Further argunents passed to \code{\link[=db_restore]{db_restore()}}.}
}
\description{
Releases are a collection of scripts and data required to update a database
from one to another version.
}
