% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_template.R
\name{copy_template}
\alias{copy_template}
\title{Insert templates to working directory}
\usage{
copy_template(filename, template, ...)
}
\arguments{
\item{filename}{A character value for the path for the copy of the template.
It may include the name of the file.}

\item{template}{A character value. The name of the template including.
Since this entry is compared by \code{\link[=pmatch]{pmatch()}}, it is not necessary to
include the extension.}

\item{...}{Further arguments passed to \code{\link[=file.copy]{file.copy()}}.}
}
\description{
The package 'backpackR' includes eine collection of templates for files
needed to be inserted into a project by \code{init_project()}.

This is a wrapper of \code{\link[=file.copy]{file.copy()}}.
}
