% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credentials.R
\name{credentials}
\alias{credentials}
\title{Prompt for credentials}
\usage{
credentials()
}
\description{
Prompt for credentials
}
\details{
A wizard that asks for credentials (username and password).
This is to avoid writing passwords in the scripts.
The credentials are then passed to other functions that require them.
}
\author{
Miguel Alvarez
}
