% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_backup.R
\name{db_backup}
\alias{db_backup}
\title{Create a backup by 'pg_dump' and restore database}
\usage{
db_backup(
  dbname,
  host = "localhost",
  port = "5432",
  path = ".",
  filename,
  fext = ".backup",
  path_psql = "/usr/bin",
  ...
)
}
\arguments{
\item{dbname}{Character value indicating the name of the target PostgreSQL
database.}

\item{host}{Character value, the host name.}

\item{port}{Integer value, the port applied for database connection.}

\item{path}{Character value inidcating the folder to place the backup. It
can also be part of 'filename'.}

\item{filename}{Character value, the name of the backup file. If not
provided,
the function guess the database's name.}

\item{fext}{Character value indicating the extension used for the backup
file, including the leading dot. The extension ".backup" is used by
default.}

\item{path_psql}{Character value indicating the system path to PostgreSQL
binaries (set default for Linux).}

\item{...}{Further arguments passed to \code{\link[=system]{system()}}.}

\item{backup}{A character value indicating a specific backup file to be used
for restore.}

\item{opts}{A character value indicating additional options for "pg_restore",
for instance "--clean" or "--create".}
}
\description{
Create a backup by 'pg_dump' and restore database
}
\details{
The function \code{db_backup()} is a wrapper for \strong{pg_dump} and is creating
writing a dump file named by the backed up database, a time stamp (date of
creation) and a suffix (in the case of a further backup done at the same
day).

The function \code{db_restore()} can restore the created backup files through
\strong{pg_restore}.
If several backups have been accumulated in a common folder, this function is
able to recognize the newest one.
}
\author{
Miguel Alvarez
}
