% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_project.R
\name{init_project}
\alias{init_project}
\title{Initialize a new project for database updates}
\usage{
init_project(path, dbname, main_script = "main-script", ...)
}
\arguments{
\item{path}{A character value indicating the path for the new project.}

\item{dbname}{A character value with the name of the database to be updated.}

\item{main_script}{A character value with the name of the template used as
main script in the project.}

\item{...}{Further arguments (not yet in use).}
}
\description{
A project folder will contain all necessary data and scripts needed to update
a database into a new version. This strategy will allow to rebuild the
database in a preferential version and document at the same time all the
process done on the way.
}
