% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_releases.R
\name{sort_releases}
\alias{sort_releases}
\title{Chronological list of releases}
\usage{
sort_releases(path, dbname, ...)
}
\arguments{
\item{path}{A character value. The path to the collection of releases.}

\item{dbname}{A character value indicating the name of the database.}

\item{...}{Further arguments (not yet in use).}
}
\description{
If a central folder is used to collect releases, a chronological list may be
required to restore a particular state of the database.
This feature is more likely to be used as an auxiliary by other functions in
the package.
}
