% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_backup.R
\name{db_backup}
\alias{db_backup}
\alias{db_restore}
\title{Create a backup by 'pg_dump' and restore database}
\usage{
db_backup(
  dbname,
  filename,
  host = "localhost",
  port = "5432",
  path_psql = "/usr/bin",
  user,
  pw,
  ...
)

db_restore(
  dbname,
  filename,
  host = "localhost",
  port = "5432",
  path_psql = "/usr/bin",
  opts = "--clean",
  user,
  pw,
  ...
)
}
\arguments{
\item{dbname}{Character value indicating the name of the target PostgreSQL
database.}

\item{filename}{Character value, the name of the backup file including the
path to location and the extension (.backup). If not provided,
the function guess the database's name.}

\item{host}{Character value, the host name.}

\item{port}{Integer value, the port applied for database connection.}

\item{path_psql}{Character value indicating the system path to PostgreSQL
binaries (set default for Linux).}

\item{user}{A character value with the name of the user of the database.
If missing it will be prompted by \code{\link[=credentials]{credentials()}}.}

\item{pw}{A character value with the password (not recommended to set in
the script). If missing it will be prompted by \code{\link[=credentials]{credentials()}}.}

\item{...}{Further arguments passed to \code{\link[=system]{system()}}.}

\item{opts}{A character value indicating additional options for "pg_restore",
for instance "--clean" or "--create".}
}
\value{
An invisible vector with the name of the database and the name of the user.
}
\description{
Create a backup by 'pg_dump' and restore database
}
\details{
The function \code{db_backup()} is a wrapper for \strong{pg_dump}.

The function \code{db_restore()} can restore the created backup files through
\strong{pg_restore}.

A prompt will request the name of the user and the respective password,
except when they are already provided in the function's call.
}
\author{
Miguel Alvarez
}
