% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_project.R
\name{init_project}
\alias{init_project}
\title{Initialize a new project for database updates}
\usage{
init_project(
  path,
  dbname,
  main_script = "main-script",
  host = "localhost",
  port = "5432",
  user = "",
  password = "",
  run_last_bk = TRUE,
  path_bk,
  ...
)
}
\arguments{
\item{path}{A character value indicating the path for the new project.}

\item{dbname}{A character value with the name of the database to be updated.}

\item{main_script}{A character value with the name of the template used as
main script in the project.}

\item{host}{Character value, the host name.}

\item{port}{Integer value, the port applied for database connection.}

\item{user}{A character value with the name of the user connecting the
database. If not provided, it will be prompted by \code{\link[=credentials]{credentials()}}.}

\item{password}{A character value with the password of the user. If not
provided, it will be prompted by \code{\link[=credentials]{credentials()}}.}

\item{run_last_bk}{A logical value whether the last release of the database
should be first restored. This will be then performed by \code{\link[=build_db]{build_db()}}.}

\item{path_bk}{A character value indicating the path to the folder containing
the collection of backup files.}

\item{...}{Further arguments (not yet in use).}
}
\description{
A project folder will contain all necessary data and scripts needed to update
a database into a new version. This strategy will allow to rebuild the
database in a preferential version and document at the same time all the
process done on the way.
}
